+++ server/src/api/pool/__init__.py
@@ -1,2 +1,3 @@
 from api.pool.routes import router
+from api.pool.websocket import websocket_router
 from api.pool.tasks import queue_next_songs
