+++ server/src/api/pool/routes.py
@@ -2,9 +2,9 @@ from logging import getLogger
 
 from fastapi import APIRouter, WebSocket
 
-from api.common.dependencies import validated_user, validated_user_from_query_parameters
-from api.pool.dependencies import PoolSpotifyClient, PoolDatabaseConnection, PoolPlaybackService, PoolWebsocketUpdater, \
-    PlaybackWebsocketUpdater
+from api.common.dependencies import validated_user
+from api.pool.dependencies import PoolSpotifyClient, PoolDatabaseConnection, PoolPlaybackService, \
+    WebsocketUpdater
 from api.pool.helpers import create_pool_return_model
 from api.pool.models import PoolCreationData, PoolFullContents, PoolContent, PoolTrack
 from database.entities import User, PoolMember
@@ -37,32 +37,33 @@ async def get_pool(user: validated_user, database_connection: PoolDatabaseConnec
 @router.post("/content")
 async def add_content(to_add: PoolContent, user: validated_user, spotify_client: PoolSpotifyClient,
                       database_connection: PoolDatabaseConnection,
-                      pool_websocket_updater: PoolWebsocketUpdater) -> PoolFullContents:
+                      websocket_updater: WebsocketUpdater) -> PoolFullContents:
     _logger.debug(f"POST /pool/content called with content {to_add} and token {user.session.user_token}")
     added_content = spotify_client.get_pool_content(user, to_add)
     whole_pool = database_connection.add_to_pool(added_content, user)
-    data_model = await _create_model_and_update_listeners(database_connection, pool_websocket_updater, user, whole_pool)
+    data_model = await _create_model_and_update_listeners(database_connection, websocket_updater, user, whole_pool)
     return data_model
 
 
 @router.delete("/content/{spotify_uri}")
 async def delete_content(spotify_uri: str, user: validated_user, database_connection: PoolDatabaseConnection,
-                         pool_websocket_updater: PoolWebsocketUpdater) -> PoolFullContents:
+                         websocket_updater: WebsocketUpdater) -> PoolFullContents:
     _logger.debug(f"DELETE /pool/content/{spotify_uri} called with token {user.session.user_token}")
     whole_pool = database_connection.delete_from_pool(spotify_uri, user)
-    data_model = await _create_model_and_update_listeners(database_connection, pool_websocket_updater, user, whole_pool)
+    data_model = await _create_model_and_update_listeners(database_connection, websocket_updater, user, whole_pool)
     return data_model
 
 
 async def _create_model_and_update_listeners(database_connection: PoolDatabaseConnection,
-                                             pool_websocket_updater: PoolWebsocketUpdater, user: User,
+                                             websocket_updater: WebsocketUpdater, user: User,
                                              whole_pool: list[PoolMember]):
     pool_users = database_connection.get_pool_users(user)
     pool = database_connection.get_pool(user)
     code = pool.share_data.code if pool.share_data is not None else None
     current_track = database_connection.get_current_track(user)
     data_model = create_pool_return_model(whole_pool, pool_users, current_track, code)
-    await pool_websocket_updater.pool_updated(data_model, pool.id)
+    user_ids = [user.spotify_id for user in pool_users]
+    await websocket_updater.push_update(user_ids, "pool", data_model.model_dump())
     return data_model
 
 
@@ -79,28 +80,10 @@ async def share_pool(user: validated_user, pool_database_connection: PoolDatabas
 
 
 @router.post("/join/{code}")
-async def join_pool(code: str, user: validated_user, pool_websocket_updater: PoolWebsocketUpdater,
+async def join_pool(code: str, user: validated_user, pool_websocket_updater: WebsocketUpdater,
                     pool_database_connection: PoolDatabaseConnection) -> PoolFullContents:
     _logger.debug(f"POST /pool/join called with code {code} and token {user.session.user_token}")
     data_model = create_pool_return_model(*pool_database_connection.join_pool(user, code))
-    pool = pool_database_connection.get_pool(user)
-    await pool_websocket_updater.pool_updated(data_model, pool.id)
+    user_ids = [user.spotify_id for user in pool_database_connection.get_pool_users(user)]
+    await pool_websocket_updater.push_update(user_ids, "pool", data_model.model_dump())
     return data_model
-
-
-@router.websocket("/register_listener")
-async def register_for_pool_updates(websocket: WebSocket, user: validated_user_from_query_parameters,
-                                    pool_database_connection: PoolDatabaseConnection,
-                                    pool_websocket_updater: PoolWebsocketUpdater):
-    await websocket.accept()
-    pool = pool_database_connection.get_pool(user)
-    pool_websocket_updater.add_socket(websocket, pool)
-
-
-@router.websocket("/playback/register_listener")
-async def register_for_playback_updates(websocket: WebSocket, user: validated_user_from_query_parameters,
-                                        pool_database_connection: PoolDatabaseConnection,
-                                        playback_websocket_updater: PlaybackWebsocketUpdater):
-    await websocket.accept()
-    playback_pool = pool_database_connection.get_pool(user)
-    playback_websocket_updater.add_socket(websocket, playback_pool)
